requirements.txt
#libraries
streamlit==1.39.0
google-generativeai==0.8.2
pandas==2.2.3
python-dotenv==1.0.1